feat(roasting): add Turn Point events and chart visualization
This commit introduces the ability to record a specific "Turn Point" (temperature and timestamp) during a roast session and visualizes it accurately on the roasting chart.

### Key Changes:

**Data Model & Logic:**
- **Turn Point Events:** Introduced a `TurnPointEvent` data class to store temperature and time (in seconds).
- **State Management:** Updated `RoastingFormState` to include `turnPoints` and added `addTurnPoint` and `removeTurnPoint` actions to the `RoastingViewModel`.
- **Chart Accuracy:** Refactored `ChartDataPoint` and `PdfExportManager` to support `Float` interval numbers. This allows the Turn Point to be plotted precisely between standard time intervals on the chart.
- **Data Sorting:** The chart logic now sorts data points by total seconds to ensure the temperature line is drawn correctly when non-interval events like Turn Points are added.

**UI Enhancements:**
- **Turn Point Configuration:** Added a "Pengaturan Turn Point" card in `RoastingFormScreen`. Users can now input a specific temperature and second for the turn point, which is displayed as an `InputChip` with delete functionality.
- **Filtered Table View:** Updated the data table below the chart to filter out non-integer interval points (like Turn Points), keeping the grid clean while still showing the data on the graph.
- **PDF Export:** Updated `PdfExportManager` to use the high-resolution chart data (including Turn Points) for the generated reports.

**UI Refinements:**
- Removed the manual "Turn Point (°C)" text input in favor of the new dedicated configuration card.
- Standardized "Simpan" buttons to "Save" for consistency across the form.

+++ app/src/main/java/com/indie/roastlog/pdf/PdfExportManager.kt
@@ -53,7 +53,7 @@ data class RoastSessionData(
     val intervalSeconds: Int,
     val burnerIntervalSeconds: Int,
     val startTemperature: Float,
-    val temperatureData: List<Pair<Int, Float>>, // intervalNumber, temperature
+    val temperatureData: List<Pair<Float, Float>>, // intervalNumber (Float), temperature
     val roastDate: Date = Date()
 )
 
@@ -355,7 +355,7 @@ class PdfExportManager(private val context: Context) {
         }
 
         val entries = data.temperatureData.map { (intervalNum, temp) ->
-            Entry(intervalNum.toFloat(), temp)
+            Entry(intervalNum, temp)
         }
 
         val dataSet = LineDataSet(entries, "Temperature").apply {
